+++ ui/components/espace_pro/Authentification/index.ts
@@ -1,5 +1,4 @@
-import CompteEnAttente from "./CompteEnAttente"
 import CreationCompte from "./CreationCompte"
 import OptOutValidation from "./OptOutValidation"
 
-export { CreationCompte, OptOutValidation, CompteEnAttente }
+export { CreationCompte, OptOutValidation }
