+++ ui/components/espace_pro/FormulaireCreationOffre/FormulaireCreationOffre.tsx
@@ -1,0 +1,240 @@
+import { Box, Grid, Heading, Text } from "@chakra-ui/react"
+import dayjs from "dayjs"
+import { Formik } from "formik"
+import { omit } from "lodash-es"
+import { useRouter } from "next/router"
+import { useContext, useState } from "react"
+import { useQuery } from "react-query"
+import { IRecruiterJson, IReferentielRomeForJob } from "shared"
+import { IJobJson, JOB_STATUS } from "shared/models/job.model"
+import * as Yup from "yup"
+
+import { AUTHTYPE } from "@/common/contants"
+import { InfosDiffusionOffre } from "@/components/DepotOffre/InfosDiffusionOffre"
+import { RomeDetailWithQuery } from "@/components/DepotOffre/RomeDetailWithQuery"
+import { WidgetContext } from "@/context/contextWidget"
+import { useAuth } from "@/context/UserContext"
+import { createOffre, createOffreByToken, getFormulaire, getFormulaireByToken, getRelatedEtablissementsFromRome, getRomeDetail } from "@/utils/api"
+
+import { FormikCreationOffreButtons } from "./FormulaireCreationOffreButtons"
+import { FormikCreationOffreFields } from "./FormulaireCreationOffreFields"
+
+const ISO_DATE_FORMAT = "YYYY-MM-DD"
+const FR_DATE_FORMAT = "DD/MM/YYYY"
+
+export const FormulaireCreationOffre = ({
+  fromDashboard,
+  offre,
+  handleSave,
+}: {
+  fromDashboard?: boolean
+  offre?: IJobJson
+  handleSave?: (values: any) => Promise<{ form?: IRecruiterJson; offre?: IJobJson }>
+}) => {
+  const { rome_appellation_label, rome_code } = offre ?? {}
+  const initRome = rome_code?.at(0)
+  const [romeAndAppellation, setRomeAndAppellation] = useState<{ rome: string; appellation: string }>(
+    rome_appellation_label && initRome ? { rome: initRome, appellation: rome_appellation_label } : null
+  )
+  const { rome } = romeAndAppellation ?? {}
+  const { user } = useAuth()
+  const { widget } = useContext(WidgetContext)
+  const router = useRouter()
+  const { establishment_id, email, userId, token } = router.query as { establishment_id: string; email: string; userId: string; type: string; token: string }
+
+  const { data: formulaire } = useQuery("offre-liste", {
+    enabled: Boolean(establishment_id),
+    queryFn: () => (token ? getFormulaireByToken(establishment_id, token) : getFormulaire(establishment_id)),
+  })
+  const romeQuery = useQuery(["getRomeDetail", rome], () => getRomeDetail(rome), {
+    retry: false,
+    enabled: Boolean(rome),
+  })
+
+  const { geo_coordinates } = formulaire ?? {}
+  const [latitude, longitude] = geo_coordinates?.split(",")?.map((str) => parseFloat(str)) ?? []
+  const relatedEtablissementQuery = useQuery(["relatedEtablissement", rome, latitude, longitude], () => getRelatedEtablissementsFromRome({ rome, latitude, longitude, limit: 1 }), {
+    retry: false,
+    enabled: Boolean(rome && latitude && longitude),
+  })
+
+  const haveProposals = Boolean(relatedEtablissementQuery.data?.length)
+
+  const {
+    widget: { isWidget },
+  } = useContext(WidgetContext)
+
+  const [selectedCompetences, setSelectedCompetences] = useState<IReferentielRomeForJob["competences"] | null>(offre?.competences_rome ?? null)
+  const [competencesDirty, setCompetencesDirty] = useState(false)
+
+  const onRomeChange = (rome: string, appellation: string) => {
+    setRomeAndAppellation({ rome, appellation })
+    setSelectedCompetences(null)
+    setCompetencesDirty(true)
+  }
+
+  const onSelectedCompetencesChange = (selectedCompetences: Record<string, string[]>) => {
+    if (!romeQuery.data) {
+      throw new Error("inattendu : pas de données ROME")
+    }
+    const { competences } = romeQuery.data as IReferentielRomeForJob
+    const isSelected = (groupKey: string, competence: string): boolean => (selectedCompetences[groupKey] ?? []).includes(competence)
+
+    const savedCompetences: IReferentielRomeForJob["competences"] = {
+      savoir_etre_professionnel: competences.savoir_etre_professionnel.filter((x) => isSelected("savoir_etre_professionnel", x.libelle)),
+      savoir_faire: competences.savoir_faire.flatMap((competencesGroup) => {
+        const selectedItems = (competencesGroup?.items ?? []).filter(({ libelle }) => isSelected("savoir_faire", libelle))
+        if (!selectedItems.length) {
+          return []
+        }
+        return [{ ...competencesGroup, items: selectedItems }]
+      }),
+      savoirs: competences.savoirs.flatMap((competencesGroup) => {
+        const selectedItems = (competencesGroup?.items ?? []).filter(({ libelle }) => isSelected("savoirs", libelle))
+        if (!selectedItems.length) {
+          return []
+        }
+        return [{ ...competencesGroup, items: selectedItems }]
+      }),
+    }
+    setSelectedCompetences(savedCompetences)
+    setCompetencesDirty(true)
+  }
+
+  if (!establishment_id) return <></>
+
+  /**
+   * @description Submits form from in a connected context.
+   * @param values
+   * @param resetForm
+   * @return {Promise<void>}
+   */
+  const submitFromDashboard = async (values, { resetForm }) => {
+    if (user && user.type !== AUTHTYPE.ENTREPRISE) {
+      await handleSave(values)
+    } else {
+      const res = await handleSave(values)
+
+      // Only redirect user in case of offer creation
+      if (res) {
+        await handleRedirectionAfterSubmit(res.form, res.offre, true, null)
+      }
+    }
+
+    resetForm({})
+  }
+
+  /**
+   * @description Submits the form.
+   * If there are some CFA proposals, user will be redirected on the CFA selection page.
+   * Otherwise, he's redirected at the end of the form.
+   * @param values
+   * @return {Promise<void>}
+   */
+  const submitFromDepotRapide = async (values) => {
+    const { recruiter: formulaire, token: jobToken } = await (token ? createOffreByToken(establishment_id, values, token) : createOffre(establishment_id, values))
+    const [job] = formulaire.jobs.slice(-1)
+    await handleRedirectionAfterSubmit(formulaire, job, false, jobToken)
+  }
+
+  const handleRedirectionAfterSubmit = (form: IRecruiterJson, job: IJobJson, fromDashboard: boolean, jobToken: string) => {
+    if (haveProposals) {
+      return router.replace({
+        pathname: isWidget ? "/espace-pro/widget/entreprise/mise-en-relation" : "/espace-pro/creation/mise-en-relation",
+        query: { job: JSON.stringify(omit(job, "rome_detail")), email, geo_coordinates: form.geo_coordinates, fromDashboard, userId, establishment_id, token: jobToken },
+      })
+    }
+
+    router.replace({
+      pathname: isWidget ? "/espace-pro/widget/entreprise/fin" : "/espace-pro/creation/fin",
+      query: { jobId: job._id.toString(), email, withDelegation: false, fromDashboard, userId, establishment_id, token: jobToken },
+    })
+  }
+
+  const finalSelectedCompetences = selectedCompetences ?? romeQuery?.data?.competences
+
+  const onSubmit = (values, bag) => {
+    values = { ...values, competences_rome: finalSelectedCompetences }
+    fromDashboard ? submitFromDashboard(values, bag) : submitFromDepotRapide(values)
+  }
+
+  const minStartDate = dayjs().startOf("day")
+  const maxStartDate = dayjs().add(2, "years")
+
+  return (
+    <>
+      {widget.isWidget && (
+        <Text textTransform="uppercase" fontSize="20px" color="#666666">
+          Dépot simplifié d'offre en alternance
+        </Text>
+      )}
+      <Formik
+        validateOnMount
+        enableReinitialize={true}
+        initialValues={{
+          rome_label: offre?.rome_label ?? "",
+          rome_appellation_label: offre?.rome_appellation_label ?? "",
+          rome_code: offre?.rome_code ?? [],
+          job_level_label: offre?.job_level_label ?? "Indifférent",
+          job_start_date: offre?.job_start_date ? dayjs(offre.job_start_date).format(ISO_DATE_FORMAT) : "",
+          job_creation_date: offre?.job_creation_date ?? dayjs().format(ISO_DATE_FORMAT),
+          job_expiration_date: offre?.job_expiration_date ?? dayjs().add(2, "month").format(ISO_DATE_FORMAT),
+          job_status: offre?.job_status ?? JOB_STATUS.ACTIVE,
+          job_type: offre?.job_type ?? ["Apprentissage"],
+          delegations: offre?.delegations ?? undefined,
+          is_disabled_elligible: offre?.is_disabled_elligible ?? false,
+          job_count: offre?.job_count ?? 1,
+          job_duration: offre?.job_duration ?? 12,
+          job_rythm: offre?.job_rythm ?? null,
+          job_delegation_count: offre?.job_delegation_count ?? 0,
+        }}
+        validationSchema={Yup.object().shape({
+          rome_label: Yup.string().required("Champ obligatoire"),
+          job_level_label: Yup.string().required("Champ obligatoire"),
+          job_start_date: Yup.date()
+            .min(minStartDate, `La date de début doit être après le ${minStartDate.format(FR_DATE_FORMAT)}`)
+            .max(maxStartDate, `La date de début doit être avant le ${maxStartDate.format(FR_DATE_FORMAT)}`)
+            .required("Champ obligatoire"),
+          job_type: Yup.array().required("Champ obligatoire"),
+          job_duration: Yup.number().max(36, "Durée maximale du contrat : 36 mois").min(6, "Durée minimale du contrat : 6 mois").typeError("Durée minimale du contrat : 6 mois"),
+        })}
+        onSubmit={(values: any, bag) => onSubmit(values, bag)}
+      >
+        {() => (
+          <Grid
+            gridTemplateColumns={["repeat(1, 1fr)", "repeat(1, 1fr)", "repeat(2, 1fr)", "repeat(2, 1fr)"]}
+            gridTemplateRows={["repeat(3, auto)", "repeat(3, auto)", "auto 1fr", "auto 1fr"]}
+            rowGap={4}
+            columnGap={4}
+          >
+            <Box>
+              <Heading className="big">Votre offre</Heading>
+              <Box mt={4}>
+                <FormikCreationOffreFields onRomeChange={onRomeChange} />
+              </Box>
+            </Box>
+            <Box gridColumnStart={[1, 1, 2, 2]} gridRow={["2 / 3", "2 / 3", "1 / 3", "1 / 3"]}>
+              {romeAndAppellation ? (
+                <RomeDetailWithQuery
+                  selectedCompetences={{
+                    savoirs: (finalSelectedCompetences?.savoirs ?? []).flatMap(({ items = [] }) => items.map((item) => item?.libelle)),
+                    savoir_etre_professionnel: (finalSelectedCompetences?.savoir_etre_professionnel ?? []).flatMap(({ libelle }) => (libelle ? [libelle] : [])),
+                    savoir_faire: (finalSelectedCompetences?.savoir_faire ?? []).flatMap(({ items = [] }) => items.map((item) => item?.libelle)),
+                  }}
+                  appellation={romeAndAppellation.appellation}
+                  rome={romeAndAppellation.rome}
+                  onChange={onSelectedCompetencesChange}
+                />
+              ) : (
+                <InfosDiffusionOffre />
+              )}
+            </Box>
+            <Box mt={8}>
+              <FormikCreationOffreButtons competencesDirty={competencesDirty} />
+            </Box>
+          </Grid>
+        )}
+      </Formik>
+    </>
+  )
+}
